Do not use route.state
Got several ' in-app'  warnings when cloning the repo, ' yarn'-ing to download dependencies and running the app on android.
They referred to this, https://reactnavigation.org/docs/route-prop/ stating that you should not use the route.state.

I looked at what it does, and basically it just seems to be checking if we are in a screen where we want to show the tabs or not. This updated code is I think equivalent.